+++ SukiUI.Demo/Features/ControlsLibrary/ExpanderViewModel.cs
@@ -1,9 +1,8 @@
 using Material.Icons;
 
-namespace SukiUI.Demo.Features.ControlsLibrary
+namespace SukiUI.Demo.Features.ControlsLibrary;
+
+public class ExpanderViewModel() : DemoPageBase("Expander", MaterialIconKind.SwapVertical)
 {
-    public class ExpanderViewModel() : DemoPageBase("Expander", MaterialIconKind.SwapVertical)
-    {
         
-    }
 }
