+++ SukiUI.Demo/Features/ControlsLibrary/TabControl/TabControlViewModel.cs
@@ -1,9 +1,8 @@
 using System.Linq;
 using Avalonia.Collections;
-using CommunityToolkit.Mvvm.ComponentModel;
 using Material.Icons;
 
-namespace SukiUI.Demo.Features.ControlsLibrary;
+namespace SukiUI.Demo.Features.ControlsLibrary.TabControl;
 
 public class TabControlViewModel : DemoPageBase
 {
@@ -14,11 +13,4 @@ public class TabControlViewModel : DemoPageBase
         Tabs.AddRange(Enumerable.Range(1, 5).Select(x => new TabViewModel(x)));
         Tabs.Add(new TabViewModel(6, false));
     }
-}
-
-public class TabViewModel(int index, bool isEnabled = true) : ObservableObject
-{
-    public string Header { get; } = $"Tab {index}";
-    public string Content { get; } = $"Tab {index} Content.";
-    public bool IsEnabled { get; } = isEnabled;
 }
