+++ SukiUI.Demo/Features/ControlsLibrary/Toasts/ToastWindowDemo.axaml.cs
@@ -1,4 +1,3 @@
-using Avalonia.Controls;
 using Avalonia.Interactivity;
 using SukiUI.Controls;
