+++ SukiUI.Demo/Features/ControlsLibrary/TogglesViewModel.cs
@@ -2,6 +2,6 @@ using Material.Icons;
 
 namespace SukiUI.Demo.Features.ControlsLibrary;
 
-public partial class TogglesViewModel() : DemoPageBase("Toggles", MaterialIconKind.ToggleSwitch)
+public class TogglesViewModel() : DemoPageBase("Toggles", MaterialIconKind.ToggleSwitch)
 {
 }
